examples/Microbiology/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.5
aiosignal==1.4.0
annotated-doc==0.0.4
annotated-types==0.7.0
anthropic==0.46.0
anyio==4.13.0
attrs==26.1.0
backoff==2.2.1
beautifulsoup4==4.14.3
brotli==1.2.0
certifi==2026.2.25
cffi==2.0.0
cfgv==3.5.0
charset-normalizer==3.4.7
click==8.3.2
cobble==0.1.4
cryptography==46.0.7
cssselect2==0.9.0
cuda-bindings==13.2.0
cuda-pathfinder==1.5.3
cuda-toolkit==13.0.2
datasets==4.8.4
dill==0.4.1
distlib==0.4.0
distro==1.9.0
dnspython==2.8.0
docstring_parser==0.17.0
EbookLib==0.18
einops==0.8.2
email-validator==2.3.0
et_xmlfile==2.0.0
fastapi==0.135.3
fastapi-cli==0.0.24
fastapi-cloud-cli==0.16.1
fastapi-openai-compat==1.2.0
fastar==0.10.0
filelock==3.25.2
filetype==1.2.0
fonttools==4.62.1
frozenlist==1.8.0
fsspec==2026.2.0
ftfy==6.3.1
google-auth==2.49.2
google-genai==1.73.1
groq==1.1.2
h11==0.16.0
hayhooks==1.16.0
haystack-ai==2.27.0
haystack-experimental==0.19.0
hf-xet==1.4.3
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
huggingface_hub==0.36.2
identify==2.6.18
idna==3.11
Jinja2==3.1.6
jiter==0.13.0
joblib==1.5.3
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
lazy_imports==1.2.0
loguru==0.7.3
lxml==6.0.4
mammoth==1.12.0
markdown-it-py==4.0.0
markdown2==2.5.5
markdownify==1.2.2
marker-pdf==1.10.2
MarkupSafe==3.0.3
mdurl==0.1.2
more-itertools==11.0.1
mpmath==1.3.0
multidict==6.7.1
multiprocess==0.70.19
networkx==3.6.1
nltk==3.9.4
nodeenv==1.10.0
numpy==2.4.4
nvidia-cublas==13.1.0.3
nvidia-cuda-cupti==13.0.85
nvidia-cuda-nvrtc==13.0.88
nvidia-cuda-runtime==13.0.96
nvidia-cudnn-cu13==9.19.0.56
nvidia-cufft==12.0.0.61
nvidia-cufile==1.15.1.6
nvidia-curand==10.4.0.35
nvidia-cusolver==12.0.4.66
nvidia-cusparse==12.6.3.3
nvidia-cusparselt-cu13==0.8.0
nvidia-nccl-cu13==2.28.9
nvidia-nvjitlink==13.0.88
nvidia-nvshmem-cu13==3.4.5
nvidia-nvtx==13.0.85
ollama==0.6.1
ollama-haystack==6.2.0
openai==1.109.1
opencv-python-headless==4.11.0.86
openpyxl==3.1.5
packaging==26.0
pandas==3.0.2
pdftext==0.6.3
pgvector==0.4.2
pgvector-haystack==6.3.0
pillow==10.4.0
platformdirs==4.9.6
posthog==7.9.12
pre_commit==4.5.1
propcache==0.4.1
psycopg==3.3.3
psycopg-binary==3.3.3
pyarrow==23.0.1
pyasn1==0.6.3
pyasn1_modules==0.4.2
pycparser==3.0
pydantic==2.12.5
pydantic-extra-types==2.11.2
pydantic-settings==2.13.1
pydantic_core==2.41.5
pydyf==0.12.1
Pygments==2.20.0
pypdf==6.9.2
pypdfium2==4.30.0
pyphen==0.17.2
python-dateutil==2.9.0.post0
python-discovery==1.2.2
python-dotenv==1.2.2
python-multipart==0.0.24
python-pptx==1.0.2
PyYAML==6.0.3
RapidFuzz==3.14.5
referencing==0.37.0
regex==2024.11.6
requests==2.33.1
rich==14.3.3
rich-toolkit==0.19.7
rignore==0.7.6
rpds-py==0.30.0
safetensors==0.7.0
scikit-learn==1.8.0
scipy==1.17.1
sentry-sdk==2.57.0
setuptools==81.0.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
soupsieve==2.8.3
starlette==1.0.0
surya-ocr==0.17.1
sympy==1.14.0
tenacity==9.1.4
threadpoolctl==3.6.0
tinycss2==1.5.1
tinyhtml5==2.1.0
tokenizers==0.22.2
torch==2.11.0
tqdm==4.67.3
transformers==4.57.6
triton==3.6.0
typer==0.24.1
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
uvicorn==0.44.0
uvloop==0.22.1
virtualenv==21.2.4
watchfiles==1.1.1
wcwidth==0.6.0
weasyprint==63.1
webencodings==0.5.1
websockets==16.0
xlsxwriter==3.2.9
xxhash==3.6.0
yarl==1.23.0
zopfli==0.4.1
